# src/main/resources/static/swagger-api/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "HaoHaoShop"
servers:
  - url: "http://localhost:8080"
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - "AuthenticationController"
      summary: "Login"
      operationId: "login"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "test123@test.com"
                password:
                  type: string
                  example: "1234567890"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
        "403":
          description: "Forbidden"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Incorrect password or email"
  /api/v1/auth/forgot:
    post:
      tags:
        - "AuthenticationController"
      summary: "Send password reset code to users email"
      operationId: "forgotPassword"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "test123@test.com"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Reset password code is send to your E-mail"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Email not found"
  /api/v1/auth/reset/{code}:
    get:
      tags:
        - "AuthenticationController"
      summary: "Returns user info by password reset code"
      operationId: "getPasswordResetCode"
      parameters:
        - name: "code"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Password reset code is invalid!"
  /api/v1/auth/reset:
    post:
      tags:
        - "AuthenticationController"
      summary: "Reset password"
      operationId: "passwordReset"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: "1234567890"
                password2:
                  type: string
                  example: "1234567890"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Password successfully changed!"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  passwordError:
                    type: "string"
                  password2Error:
                    type: "string"
              examples:
                PasswordMatchError:
                  summary: "Password Match Error"
                  value:
                    passwordError: "Passwords do not match."
                PasswordConfirmationError:
                  summary: "Password Confirmation Error"
                  value:
                    password2Error: "Password confirmation cannot be empty."
...
  /api/v1/perfumes/{id}:
    get:
      tags:
        - "PerfumeController"
      summary: "Get perfume by id"
      operationId: "getPerfume"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/ids:
    post:
      tags:
        - "PerfumeController"
      summary: "Get perfumes by ids"
      operationId: "getPerfumesByIds"
      requestBody:
        content:
          'application/json':
            schema:
              type: "array"
              items:
                type: "integer"
                format: int64
              example: [1, 2, 3]
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/search:
    post:
      tags:
        - "PerfumeController"
      summary: "Find perfumes by filter parameters"
      operationId: "findPerfumesByFilterParams"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                perfumers:
                  type: "array"
                  items:
                    type: "string"
                    example: "Dior"
                genders:
                  type: "array"
                  items:
                    type: "string"
                    example: "male"
                prices:
                  type: "array"
                  items:
                    type: "integer"
                    format: int32
                  example: [1, 200]
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/search/gender:
    post:
      tags:
        - "PerfumeController"
      summary: "Find perfumes by gender"
      operationId: "findByPerfumeGender"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                genders:
                  type: "array"
                  items:
                    type: "string"
                    example: "male"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/search/perfumer:
    post:
      tags:
        - "PerfumeController"
      summary: "Find perfumes by brand"
      operationId: "findByPerfumer"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                perfumers:
                  type: "array"
                  items:
                    type: "string"
                  example: ["Dior", "Creed"]
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/graphql/ids:
    post:
      tags:
        - "PerfumeController"
      summary: "Get list of perfumes by ids GraphQL query"
      operationId: "getPerfumesByIdsQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ perfumesIds(ids: [16,17,18]) {id perfumeTitle perfumer price filename} }"

      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: "integer"
                      format: int64
                    perfumeTitle:
                      type: "string"
                    perfumer:
                      type: "string"
                    price:
                      type: "integer"
                      format: int32
                    filename:
                      type: "string"
                  example:
                    id: 43
                    perfumeTitle: "Sauvage"
                    perfumer: "Dior"
                    price: 62
                    filename: "empty.jpg"
  /api/v1/perfumes/graphql/perfumes:
    post:
      tags:
        - "PerfumeController"
      summary: "Get list of perfumes by GraphQL query"
      operationId: "getAllPerfumesByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ perfumes {id perfumeTitle perfumer price filename} }"

      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: "integer"
                      format: int64
                    perfumeTitle:
                      type: "string"
                    perfumer:
                      type: "string"
                    price:
                      type: "integer"
                      format: int32
                    filename:
                      type: "string"
                  example:
                    id: 43
                    perfumeTitle: "Sauvage"
                    perfumer: "Dior"
                    price: 62
                    filename: "empty.jpg"
  /api/v1/perfumes/graphql/perfume:
    post:
      tags:
        - "PerfumeController"
      summary: "Get perfume by GraphQL query"
      operationId: "getPerfumeByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ perfume(id: 43) {id perfumeTitle perfumer price filename} }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  id:
                    type: "integer"
                    format: int64
                  perfumeTitle:
                    type: "string"
                  perfumer:
                    type: "string"
                  price:
                    type: "integer"
                    format: int32
                  filename:
                    type: "string"
                example:
                  id: 43
                  perfumeTitle: "Sauvage"
                  perfumer: "Dior"
                  price: 62
                  filename: "empty.jpg"
...
  /api/v1/users/edit/password:
    put:
      tags:
        - "UserController"
      summary: "Change user password"
      operationId: "updateUserPassword"
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                password:
                  type: "string"
                  example: "1234567890"
                password2:
                  type: "string"
                  example: "1234567890"
              required:
                - password
                - password2
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Password successfully changed!"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  passwordError:
                    type: "string"
                  password2Error:
                    type: "string"
              examples:
                BindingResultError:
                  summary: "Binding Result Errors"
                  value:
                    passwordError: "The password must be between 6 and 16 characters long"
                    password2Error: "The password confirmation must be between 6 and 16 characters long"
                MatchPasswordError:
                  summary: "Match Password Error"
                  value:
                    data: "Passwords do not match."
      security:
        - token: []
...
components:
  securitySchemes:
    token:
      type: apiKey
      name: Authorization
      in: header
  schemas:
    PerfumeResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 43
        perfumeTitle:
          type: "string"
          example: "Sauvage"
        perfumer:
          type: "string"
          example: "Dior"
        year:
          type: "integer"
          format: "int32"
          example: 2015
        country:
          type: "string"
          example: "France"
        perfumeGender:
          type: "string"
          example: "male"
        fragranceTopNotes:
          type: "string"
          example: "Sichuan Pepper"
        fragranceMiddleNotes:
          type: "string"
          example: "Bergamot"
        fragranceBaseNotes:
          type: "string"
          example: "Ambroxan"
        description:
          type: "string"
          example: "Perfume description"
        filename:
          type: "string"
          example: "Dior_Sauvage.jpg"
        price:
          type: "integer"
          format: "int32"
          example: 62
        volume:
          type: "string"
          example: "100"
        type:
          type: "string"
          example: "Eau de parfum"
        reviews:
          type: "array"
          items:
            $ref: "#/components/schemas/ReviewResponse"
        file:
          type: "string"
          format: binary
    ReviewResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        author:
          type: "string"
          example: "John Doe"
        message:
          type: "string"
          example: "Message"
        date:
          type: "string"
          example: "2021-04-07"
...
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 2
        email:
          type: "string"
          example: "test123@test.com"
        firstName:
          type: "string"
          example: "John"
        lastName:
          type: "string"
          example: "Doe"
        city:
          type: "string"
          example: "New York"
        address:
          type: "string"
          example: "Wall Street 1"
        phoneNumber:
          type: "string"
          example: "123456"
        postIndex:
          type: "string"
          example: "123456"
        provider:
          type: "string"
          example: "LOCAL"
        active:
          type: "boolean"
          example: true
        activationCode:
          type: "string"
          example: ""
        passwordResetCode:
          type: "string"
          example: ""
        roles:
          type: "array"
          items:
            type: "string"
            example: "USER"
    AuthenticationResponse:
      type: "object"
      properties:
        email:
          type: "string"
        token:
          type: "string"
        userRole:
          type: "string"